Remove unused ResultsProviding protocol
